Move motion control state code to dedicated module
